MC Algorithm

+++ CodeStructureIdea/data/particle.cpp
@@ -1,5 +1,5 @@
 particle::particle() : x_(0.0), y_(0.0), z_(0.0) {}
-particle::particle(double x, double y, double z) : x_(x), y_(y), z_(z) {}
+particle::particle(float x, float y, float z) : x_(x), y_(y), z_(z) {}
 
 particle::particle(const particle& p) : x_(p.x()), y_(p.y()), z_(p.z()) {}
 
@@ -9,19 +9,19 @@ particle::particle(const particle& p) : x_(p.x()), y_(p.y()), z_(p.z()) {}
     z_ = p.z();
 }*/
 
-double const& particle::x() const {
+float const& particle::x() const {
     return x_;
 }
 
-double const& particle::y() const {
+float const& particle::y() const {
     return y_;
 }
 
-double const& particle::z() const {
+float const& particle::z() const {
     return z_;
 }
 
-std::array<double, 3> particle::getArray() const {
+std::array<float, 3> particle::getArray() const {
     return {x_, y_, z_};
 }
